% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{cluster_analysis_sub}
\alias{cluster_analysis_sub}
\title{cluster_analysis_sub}
\usage{
cluster_analysis_sub(
  raw_counts,
  resolution,
  neighbors,
  max_dimension,
  name_cluster
)
}
\arguments{
\item{raw_counts}{Raw count matrix (n_genes X n_cells).}

\item{resolution}{Numeric value specifying the parameter \emph{resolution}
used in the Seurat function \emph{FindClusters}.}

\item{neighbors}{Numeric value specifying the parameter \emph{k.param} in
the Seurat function \emph{FindNeighbors}}

\item{max_dimension}{Numeric value specifying the maximum number of the PCA
dimensions used in the parameter \emph{dims} for the Seurat function
\emph{FindNeighbors}}

\item{name_cluster}{Character.Name of the original cluster for which the sub
clustering is done.}
}
\value{
Seurat object including raw and normalized counts matrices and cluster result.
}
\description{
cluster_analysis_sub
}
\seealso{
\url{https://www.rdocumentation.org/packages/Seurat/versions/4.0.1/topics/RunPCA}
\url{https://www.rdocumentation.org/packages/Seurat/versions/4.0.1/topics/FindVariableFeatures}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
