% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{detect_localized_genes}
\alias{detect_localized_genes}
\title{detect_localized_genes}
\usage{
detect_localized_genes(
  knn_matrix,
  norm_matrix,
  genes_sort,
  top_number,
  threshold = 1
)
}
\arguments{
\item{knn_matrix}{K-nearest neighbors matrix (n_cells X n_cells).}

\item{norm_matrix}{Norm count matrix (n_genes X n_cells).}

\item{genes_sort}{vector of genes names sorted according to the p value given by the output of 
\emph{CIARA}.}

\item{top_number}{Number of \emph{genes_sort} to consider.}

\item{threshold}{threshold in expression for a given gene}
}
\value{
List with two elements. 
\item{list_intersect}{The first element is a list with length equal to the number of cells in \emph{norm_matrix}. 
Each entry contains the name of \emph{genes_sort} expressed in at least two neighboring cells.}
\item{rank_intersect}{The second element is a vector with length equal to the number of cells in \emph{norm_matrix}.Each entry contains the number 
of \emph{genes_sort} expressed in at least two neighboring cells.}
}
\description{
detect_localized_genes
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
