% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{find_resolution}
\alias{find_resolution}
\title{find_resolution}
\usage{
find_resolution(seurat_object, resolution_vector)
}
\arguments{
\item{seurat_object}{Seurat object as returned by
\emph{cluster_analysis_integrate_rare}}

\item{resolution_vector}{vector with all values of resolution for which the
Seurat function \emph{FindClusters} is run}
}
\value{
Clustree object showing the connection between clusters obtained at different level of resolution as specified in \emph{resolution_vector}.
}
\description{
find_resolution
}
\seealso{
\url{https://CRAN.R-project.org/package=clustree}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
