% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_functions.R
\name{merge_cluster}
\alias{merge_cluster}
\title{merge_cluster}
\usage{
merge_cluster(old_cluster, new_cluster, max_number = NULL)
}
\arguments{
\item{old_cluster}{original cluster asignment that need to be updated}

\item{new_cluster}{new cluster asignment that need to be integrated with
\emph{old_cluster}.}

\item{max_number}{Threshold in size for clusters in \emph{new_cluster}. Only
cluster with number of cells smaller than \emph{max_number} will be
integrated in \emph{old cluster}. If \emph{max_number} is NULL, then the final partition is equal to \emph{new_cluster}.}
}
\description{
merge_cluster
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
