% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{names_localized_genes}
\alias{names_localized_genes}
\title{names_localized_genes}
\usage{
names_localized_genes(list_intersect, genes_sort, max_number = 5)
}
\arguments{
\item{list_intersect}{first element of the list given as output from \emph{detect_localized_genes}.}

\item{genes_sort}{vector of genes names sorted according to the p value given by the output of 
\emph{CIARA}.}

\item{max_number}{maximum number of genes in \emph{list_intersect} to show for each cell.}
}
\value{
A vector with length equal to \emph{list_intersect}. Each entry contains the top \emph{max_number} names of genes from \emph{list_intersect}.
}
\description{
names_localized_genes
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
