% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_balons_marker}
\alias{plot_balons_marker}
\title{plot_balons_marker}
\usage{
plot_balons_marker(
  norm_counts,
  cluster,
  marker_complete,
  max_number,
  max_size = 5,
  text_size = 7
)
}
\arguments{
\item{norm_counts}{Norm count matrix (genes X cells).}

\item{cluster}{Vector of length equal to the number of cells, with cluster
assignment.}

\item{marker_complete}{Third element of the output list as
returned by the function \emph{markers_cluster_seurat}}

\item{max_number}{Integer. Maximum number of markers for each cluster for
which we want to plot the expression.}

\item{max_size}{Integer. Size of the dots to be plotted.}

\item{text_size}{Size of the text in the heatmap plot.}
}
\value{
ggplot object.
}
\description{
plot_balons_marker
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
