% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_gene}
\alias{plot_gene}
\title{plot_gene}
\usage{
plot_gene(norm_counts, coordinate_umap, gene_id, title_name)
}
\arguments{
\item{norm_counts}{Norm count matrix (genes X cells).}

\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{gene_id}{Character name of the gene.}

\item{title_name}{Character name.}
}
\value{
ggplot2 object.
}
\description{
Cells are coloured according to the expression of \emph{gene_id} and plotted
according to \emph{coordinate_umap}.
}
\seealso{
\url{https://CRAN.R-project.org/package=ggplot2}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
