% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_genes_sum}
\alias{plot_genes_sum}
\title{plot_genes_sum}
\usage{
plot_genes_sum(coordinate_umap, norm_counts, genes_relevant, name_title)
}
\arguments{
\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{norm_counts}{Norm count matrix (genes X cells).}

\item{genes_relevant}{Vector with gene names for which we want to visualize
the sum in each cell.}

\item{name_title}{Character value.}
}
\value{
ggplot2 object.
}
\description{
The sum of each gene in \emph{genes_relevant} across all cells is first
normalized to 1. Then for each cell, the sum from the (normalized) genes
expression is computed and shown in the output plot.
}
\seealso{
\url{ https://CRAN.R-project.org/package=ggplot2}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
