% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_heatmap_marker}
\alias{plot_heatmap_marker}
\title{plot_heatmap_marker}
\usage{
plot_heatmap_marker(
  marker_top,
  marker_all_cluster,
  cluster,
  condition,
  norm_counts,
  text_size
)
}
\arguments{
\item{marker_top}{First element returned by \emph{markers_cluster_seurat}}

\item{marker_all_cluster}{Second element returned by
\emph{markers_cluster_seurat}}

\item{cluster}{Vector of length equal to the number of cells, with cluster
assignment.}

\item{condition}{Vector or length equal to the number of cells, specifying
the condition of the cells (i.e. batch, dataset of origin..)}

\item{norm_counts}{Norm count matrix (genes X cells).}

\item{text_size}{Size of the text in the heatmap plot.}
}
\value{
Heatmap class object.
}
\description{
plot_heatmap_marker
}
\seealso{
\url{https://www.rdocumentation.org/packages/ComplexHeatmap/versions/1.10.2/topics/Heatmap}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
