% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_interactive}
\alias{plot_interactive}
\title{plot_interactive}
\usage{
plot_interactive(
  coordinate_umap,
  color,
  text,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL
)
}
\arguments{
\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{color}{vector of length equal to n_rows in coordinate_umap.Each cell
will be coloured following a gradient according to the corresponding value
of this vector.}

\item{text}{Character vector specifying the highly localized genes in each
cell. It is the output from \emph{selection_localized_genes}.}

\item{min_x}{Set the min limit on the x axis.}

\item{max_x}{Set the max limit on the x axis.}

\item{min_y}{Set the min limit on the y axis.}

\item{max_y}{Set the min limit on the y axis.}
}
\value{
plotly object given by \emph{plot_ly function} (from library \emph{plotly}).
}
\description{
It shows in an interactive plot which are the highly localized genes in each
cell. It is based on plotly library
}
\seealso{
\url{https://plotly.com/r/}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
