% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_localized_genes}
\alias{plot_localized_genes}
\title{plot_localized_genes}
\usage{
plot_localized_genes(
  coordinate_umap,
  norm_matrix,
  rank_intersect,
  name_title,
  colors
)
}
\arguments{
\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{norm_matrix}{Norm count matrix (n_genes X n_cells).}

\item{rank_intersect}{second element of the list given as output from \emph{detect_localized_genes}.}

\item{name_title}{Character value.}

\item{colors}{\emph{colours} parameter for the function scale_colour_gradientn from R package \emph{ggplot2}.
the sum in each cell.}
}
\value{
ggplot2 object.
}
\description{
plot_localized_genes
}
\seealso{
\url{ https://CRAN.R-project.org/package=ggplot2}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
