% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_localized_genes_interactive}
\alias{plot_localized_genes_interactive}
\title{plot_localized_genes_interactive}
\usage{
plot_localized_genes_interactive(
  coordinate_umap,
  norm_matrix,
  rank_intersect,
  text,
  ramp_list,
  min_x = NULL,
  max_x = NULL,
  min_y = NULL,
  max_y = NULL
)
}
\arguments{
\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{norm_matrix}{Norm count matrix (n_genes X n_cells).}

\item{rank_intersect}{second element of the list given as output from \emph{detect_localized_genes}.}

\item{text}{Character vector specifying the highly localized genes in each
cell. It is the output from \emph{names_localized_genes}.}

\item{ramp_list}{\emph{colors} parameter for the funnction \emph{plot_ly} from R package \emph{plotly}.}

\item{min_x}{Set the min limit on the x axis.}

\item{max_x}{Set the max limit on the x axis.}

\item{min_y}{Set the min limit on the y axis.}

\item{max_y}{Set the min limit on the y axis.}
}
\value{
plotly object given by \emph{plot_ly function} (from library \emph{plotly}).
}
\description{
It shows in an interactive plot which are the highly localized genes in each
cell. It is based on plotly library
}
\seealso{
\url{https://plotly.com/r/}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
