% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_umap}
\alias{plot_umap}
\title{plot_umap}
\usage{
plot_umap(coordinate_umap, cluster)
}
\arguments{
\item{coordinate_umap}{Data frame with dimensionality reduction coordinates.
Number of rows must be equal to the number of cells}

\item{cluster}{Vector of length equal to the number of cells, with cluster
assignment.}
}
\value{
ggplot2 object.
}
\description{
plot_umap
}
\seealso{
\url{https://CRAN.R-project.org/package=ggplot2}
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
