% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{selection_localized_genes}
\alias{selection_localized_genes}
\title{selection_localized_genes}
\usage{
selection_localized_genes(
  norm_counts,
  localized_genes,
  min_number_cells = 4,
  max_number_genes = 10
)
}
\arguments{
\item{norm_counts}{Norm count matrix (genes X cells).}

\item{localized_genes}{vector of highly localized genes as provided by the
last element of the list given as output from \emph{CIARA_mixing_final}.}

\item{min_number_cells}{Minimum number of cells where a genes must be
expressed (> 0).}

\item{max_number_genes}{Maximum number of genes to show for each cell in the
interactive plot from \emph{plot_interactive}.}
}
\value{
Character vector where each entry contains the name of the top
\emph{max_number_genes} for the corresponding cell.
}
\description{
selection_localized_genes
}
\author{
Gabriele Lubatti \email{gabriele.lubatti@helmholtz-muenchen.de}
}
